backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# OAuth
httpx==0.25.2
authlib==1.2.1

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Firebase for push notifications
firebase-admin==6.2.0

# Utils
python-dotenv==1.0.0

# Redis (optional, for caching)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
